Fix --use_on_demand flag and add test
Summary:
--use_on_demand's previous impl was not equivalent to ON_DEMAND connection type: it would try to connect to IPUs before compilation rather than after.
This is because it wasn't overriding the config's device connection type, so the program still thought we were using ALWAYS in some places.
This diff changes it so that the flag makes us override the connection type in our config to ON_DEMAND at configuration time. The rest is free.
Done at C++ level so users of the kernel directly can also use the flag.
This means the flag will now make TF connect to IPUs after compilation.

Ref T69999

Test Plan: Added test to check our internal config is correct when not using and using the flag.

Reviewers: #tensorflow, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, krzysztofk, gauthamg, mateuszk

Reviewed By: #tensorflow, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, krzysztofk, gauthamg, mateuszk

Subscribers: harrym

Maniphest Tasks: T69999

Differential Revision: https://phabricator.sourcevertex.net/D80874